Few sounds and success fixes
> The float values in the progress bars are converted into int values for a better visibility
> Coin sound removed when the game is ended
> Added a sound to signal a success completed

+++ Assets/Scripts/Menu/SuccessDescription.cs
@@ -105,7 +105,7 @@ public class SuccessDescription : MonoBehaviour,
         }
     }
 
-    public void UpdateSlider(float value)
+    public void UpdateSlider(int value)
     {
         slider.value = value;
         textDescription[2].text = value + " / " + slider.maxValue;
